# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/run-and-analyze": {
            "post": {
                "description": "This endpoint runs and analyzes the input provided in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Run and analyze the provided input",
                "parameters": [
                    {
                        "description": "Input data for analysis",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully run and analyzed the data",
                        "schema": {
                            "$ref": "#/definitions/dto.ExecuteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/runtimes": {
            "get": {
                "description": "Fetches a list of all available runtimes from the usecase layer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtimes"
                ],
                "summary": "Get all available runtimes",
                "responses": {
                    "200": {
                        "description": "List of runtimes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RuntimeResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch runtimes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health/ping": {
            "get": {
                "description": "Simple health check to see if the server is running.",
                "tags": [
                    "health"
                ],
                "summary": "Ping the server",
                "responses": {
                    "200": {
                        "description": "Pong response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ExecuteRequest": {
            "description": "Execute a code with specified parameters.",
            "type": "object",
            "required": [
                "files",
                "language",
                "version"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "compile_cpu_time": {
                    "type": "integer",
                    "default": 10000
                },
                "compile_memory_limit": {
                    "type": "integer",
                    "default": -1
                },
                "compile_timeout": {
                    "type": "integer",
                    "default": 10000
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.File"
                    }
                },
                "language": {
                    "type": "string"
                },
                "run_cpu_time": {
                    "type": "integer",
                    "default": 3000
                },
                "run_memory_limit": {
                    "type": "integer",
                    "default": -1
                },
                "run_timeout": {
                    "type": "integer",
                    "default": 3000
                },
                "stdin": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.ExecuteResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "The language used for execution.",
                    "type": "string"
                },
                "run": {
                    "description": "The result of the execution.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.Run"
                        }
                    ]
                },
                "version": {
                    "description": "The version of the language used.",
                    "type": "string"
                }
            }
        },
        "dto.File": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "encoding": {
                    "type": "string",
                    "default": "utf8"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.Run": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The exit code of the program.",
                    "type": "integer"
                },
                "cpu_time": {
                    "description": "CPU time spent in execution.",
                    "type": "integer"
                },
                "memory": {
                    "description": "Memory used by the program.",
                    "type": "integer"
                },
                "message": {
                    "description": "Message related to the execution result.",
                    "type": "string"
                },
                "output": {
                    "description": "The final output.",
                    "type": "string"
                },
                "signal": {
                    "description": "The signal received during execution.",
                    "type": "string"
                },
                "status": {
                    "description": "The status of the execution.",
                    "type": "string"
                },
                "stderr": {
                    "description": "The error output to stderr.",
                    "type": "string"
                },
                "stdout": {
                    "description": "The output to stdout.",
                    "type": "string"
                },
                "wall_time": {
                    "description": "Wall time taken for execution.",
                    "type": "integer"
                }
            }
        },
        "dto.RuntimeResponse": {
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "language": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}